Change the landing page to be ACM's page, not EECS's.
Since ACM actually has some copy written for it, and we own the domain
cwru-acm.org, it makes sense to go to the ACM page by default instead of
the EECS cabal page.

This commit should be reverted if we arrive at a more-general domain
name. And, if we point multiple domains at the same install we should
reevaluate how these routes work -- but this is just a quick fix.